+++ RELEASE-NOTES.md
@@ -23,6 +23,8 @@ Version: 1.8.2
     so common a use case among CL utils & `sysUt` is for such. (Nim code should
     "get to the point" quickly, IMO.)  Call site adaptation is pending/WIP.
 
+  - Deprecate `sysUt.findUO` since `strutils.find` has long used `memchr`
+
 Version: 1.8.1
 --------------
   - Improve default --help-syntax output